pyproject.toml
[project]
name = "idea-crucible"
version = "0.1.0"
description = "Redline-first, logic-based feasibility evaluator for startup ideas"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Idea-Crucible" }
]
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
llm = [
  "openai>=1.40.0",
]
claude = [
  "claude-agent-sdk>=0.2.0",
]

[tool.uv]
default-groups = ["main"]

